requirement.txt
python-dotenv
groq
AppOpener
pywhatkit
bs4
pillow
rich
requests
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
yahoo_fin 
beautifulsoup4 
requests 
googlesearch-python 
dotenv
speedtest-cli
googletrans==4.0.0-rc1
google_translator
google_trans_new
yfinance
watchdog
imageio imageio-ffmpeg
cmake
dlib
face_recognition